move generate reactions script to tools folder
